# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для расчета зарплаты, управления бонусами и изменениями оклада.",
        "title": "Salary Calculator API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/bonuses/": {
            "get": {
                "description": "Возвращает список всех зарегистрированных бонусов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bonuses"
                ],
                "summary": "Список бонусов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/get_bonuses.Out"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет данные существующего бонуса",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bonuses"
                ],
                "summary": "Редактировать бонус",
                "parameters": [
                    {
                        "description": "Данные бонуса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/edit_bonus.In"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/edit_bonus.Out"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новую запись о бонусе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bonuses"
                ],
                "summary": "Добавить бонус",
                "parameters": [
                    {
                        "description": "Данные бонуса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/add_bonus.In"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/add_bonus.Out"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет запись о бонусе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bonuses"
                ],
                "summary": "Удалить бонус",
                "parameters": [
                    {
                        "description": "ID бонуса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delete_bonus.In"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/delete_bonus.Out"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health/health": {
            "get": {
                "description": "Возвращает текущее состояние сервиса и системные метрики",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Проверка здоровья",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/health.Response"
                        }
                    }
                }
            }
        },
        "/salary/changes/": {
            "get": {
                "description": "Возвращает историю изменений базовой зарплаты",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "salary"
                ],
                "summary": "Список изменений зарплаты",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/get_salary_changes.Out"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет данные существующей записи об изменении оклада",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "salary"
                ],
                "summary": "Редактировать изменение зарплаты",
                "parameters": [
                    {
                        "description": "Данные изменения",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/edit_salary_change.In"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/edit_salary_change.Out"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новую запись об изменении оклада",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "salary"
                ],
                "summary": "Добавить изменение зарплаты",
                "parameters": [
                    {
                        "description": "Данные изменения",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/add_salay_change.In"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/add_salay_change.Out"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет запись об изменении оклада",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "salary"
                ],
                "summary": "Удалить изменение зарплаты",
                "parameters": [
                    {
                        "description": "ID записи",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delete_salary_change.In"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/delete_salary_change.Out"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/salary/report": {
            "get": {
                "description": "Рассчитывает зарплату за указанный месяц и год",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "salary"
                ],
                "summary": "Отчет по зарплате",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Год (напр. 2024)",
                        "name": "year",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Месяц (1-12)",
                        "name": "month",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/get_salary_report.Out"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "add_bonus.In": {
            "type": "object",
            "properties": {
                "coefficient": {
                    "type": "number"
                },
                "date": {
                    "$ref": "#/definitions/value_objects.SalaryDate"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "add_bonus.Out": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "add_salay_change.In": {
            "type": "object",
            "properties": {
                "date": {
                    "$ref": "#/definitions/value_objects.SalaryDate"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "add_salay_change.Out": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "calculator.SalaryCalculationResult": {
            "type": "object",
            "properties": {
                "advance": {
                    "type": "number"
                },
                "extra_payments": {
                    "$ref": "#/definitions/value_objects.ExtraPaymentsCollectionDto"
                },
                "grossAdvance": {
                    "type": "number"
                },
                "grossSalary": {
                    "type": "number"
                },
                "grossTotal": {
                    "type": "number"
                },
                "salary": {
                    "type": "number"
                },
                "total": {
                    "type": "number"
                }
            }
        },
        "delete_bonus.In": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "delete_bonus.Out": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "delete_salary_change.In": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "delete_salary_change.Out": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "edit_bonus.In": {
            "type": "object",
            "properties": {
                "coefficient": {
                    "type": "number"
                },
                "date": {
                    "$ref": "#/definitions/value_objects.SalaryDate"
                },
                "id": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "edit_bonus.Out": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "edit_salary_change.In": {
            "type": "object",
            "properties": {
                "date": {
                    "$ref": "#/definitions/value_objects.SalaryDate"
                },
                "id": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "edit_salary_change.Out": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "get_bonuses.Bonus": {
            "type": "object",
            "properties": {
                "coefficient": {
                    "type": "number"
                },
                "date": {
                    "$ref": "#/definitions/value_objects.SalaryDate"
                },
                "id": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "get_bonuses.Out": {
            "type": "object",
            "properties": {
                "bonuses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/get_bonuses.Bonus"
                    }
                }
            }
        },
        "get_salary_changes.Change": {
            "type": "object",
            "properties": {
                "date": {
                    "$ref": "#/definitions/value_objects.SalaryDate"
                },
                "id": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "get_salary_changes.Out": {
            "type": "object",
            "properties": {
                "changes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/get_salary_changes.Change"
                    }
                }
            }
        },
        "get_salary_report.Out": {
            "type": "object",
            "properties": {
                "base_salary": {
                    "type": "number"
                },
                "result": {
                    "$ref": "#/definitions/calculator.SalaryCalculationResult"
                }
            }
        },
        "health.MemoryInfo": {
            "type": "object",
            "properties": {
                "alloc": {
                    "type": "string"
                },
                "frees": {
                    "type": "integer"
                },
                "gc_cycles": {
                    "type": "integer"
                },
                "gc_sys": {
                    "type": "string"
                },
                "heap_alloc": {
                    "type": "string"
                },
                "heap_idle": {
                    "type": "string"
                },
                "heap_inuse": {
                    "type": "string"
                },
                "heap_objects": {
                    "type": "integer"
                },
                "heap_released": {
                    "type": "string"
                },
                "heap_sys": {
                    "type": "string"
                },
                "last_gc_timestamp": {
                    "type": "integer"
                },
                "lookups": {
                    "type": "integer"
                },
                "mallocs": {
                    "type": "integer"
                },
                "next_gc": {
                    "type": "string"
                },
                "stack_inuse": {
                    "type": "string"
                },
                "stack_sys": {
                    "type": "string"
                },
                "sys": {
                    "type": "string"
                },
                "total_alloc": {
                    "type": "string"
                }
            }
        },
        "health.Response": {
            "type": "object",
            "properties": {
                "memory": {
                    "$ref": "#/definitions/health.MemoryInfo"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/health.Status"
                    }
                },
                "status": {
                    "description": "\"healthy\", \"unhealthy\"",
                    "type": "string"
                },
                "system": {
                    "$ref": "#/definitions/health.SystemInfo"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "health.Status": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "description": "\"healthy\", \"unhealthy\"",
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "health.SystemInfo": {
            "type": "object",
            "properties": {
                "cpu_cores": {
                    "type": "integer"
                },
                "go_version": {
                    "type": "string"
                },
                "goroutines": {
                    "type": "integer"
                },
                "uptime": {
                    "type": "string"
                }
            }
        },
        "value_objects.ExtraPayment": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/value_objects.ExtraPaymentType"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "value_objects.ExtraPaymentType": {
            "type": "string",
            "enum": [
                "advance",
                "salary"
            ],
            "x-enum-varnames": [
                "Advance",
                "Salary"
            ]
        },
        "value_objects.ExtraPaymentsCollectionDto": {
            "type": "object",
            "properties": {
                "payments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/value_objects.ExtraPayment"
                    }
                },
                "total": {
                    "type": "number"
                }
            }
        },
        "value_objects.SalaryDate": {
            "type": "object"
        }
    }
}